+++ src/app/server/init-server.ts
@@ -35,6 +35,7 @@ import {
   getWebhookNodeServiceEntry
 } from '../../queued-subscription-server/get-queued-subscription-request-node-service-entry';
 import { buildWebhookResolver } from '../../queued-subscription-server';
+import { Oid } from '@rumbleship/oid';
 
 export let globalGraphQlSchema: GraphQLSchema | undefined;
 
@@ -233,7 +234,10 @@ export async function initServer(
         : // RumbleshipContextControl.getContextFrom(request)
           // subscription
           ctx.connection.context.rumbleship_context;
-
+      const user = rumbleship_context.authorizer.getUser();
+      rumbleship_context.beeline.addTraceContext({
+        user: { id: user, scope: new Oid(user).unwrap().scope }
+      });
       // To consider: plugin that attaches this doesn't build the rumbleship_context for all routes
       // but Apollo seems to trigger this `context` generation call for some (all?) of them.
       // e.g. `GET /graphql` triggers this, but we don't build our context.